fix: add public.types.ts
