+++ src/entities/Post/PostCard.tsx
@@ -4,37 +4,8 @@ import { Description } from '@/entities/Post/Description'
 import { Slider } from '@/entities/Slider/Slider'
 import { TimePublish } from '@/entities/TimePublish/TimePublish'
 import UserAvatar from '@/entities/UserAvatar/UserAvatar'
+import {PostCardProps} from '@/shared/types/public.types';
 
-export type Post = {
-  avatarOwner: string | undefined
-  createdAt: string
-  description: string | undefined
-  id: number
-  images: Images[]
-  likesCount: number
-  location: string | undefined
-  owner: Owner
-  ownerId: number
-  updatedAt: string
-  userName: string
-}
-export type Images = {
-  createdAt: string
-  fileSize: number
-  height: number
-  uploadId: string
-  url: string
-  width: number
-}
-export type Owner = {
-  firstname: string
-  lastname: string
-}
-
-type PostCardProps = {
-  openModal: (post: Post) => void
-  post: Post
-}
 
 export const PostCard = ({ openModal, post }: PostCardProps) => {
   return (
