+++ src/entities/Post/PostModal.tsx
@@ -1,15 +1,12 @@
 import React from 'react'
 
-import { Post } from '@/entities/Post/PostCard'
 import { Slider } from '@/entities/Slider/Slider'
 import { TimePublish } from '@/entities/TimePublish/TimePublish'
 import UserAvatar from '@/entities/UserAvatar/UserAvatar'
 import { Modal } from '@/shared/ui/Modal/Modal'
+import {PostModalProps} from '@/shared/types/public.types';
 
-type PostModalProps = {
-  onClose: () => void
-  post: Post
-}
+
 const PostModal = ({ onClose, post }: PostModalProps) => {
   return (
     <Modal onClose={onClose} withOutHeader>
